refactor(anyharness): promote artifacts into product domain
