+++ anyharness/crates/anyharness-lib/src/domains/artifacts/mod.rs
@@ -1,0 +1,5 @@
+pub mod manifest;
+pub mod model;
+pub mod protection;
+pub mod runtime;
+pub mod service;
